Make support for resolving list of sources from ValueSourceResolver
