+++ app/questionnaire/value_source_resolver.py
@@ -26,6 +26,27 @@ class ValueSourceResolver:
     use_default_answer: bool = False
     escape_answer_value: bool = False
 
+    def _is_answer_on_path(self, answer_id: str) -> bool:
+        if self.routing_path_block_ids:
+            block = self.schema.get_block_for_answer_id(answer_id)
+            return block is not None and block["id"] in self.routing_path_block_ids
+
+        return True
+
+    def _get_answer_value(
+        self, answer_id: str, list_item_id: Optional[str]
+    ) -> Optional[AnswerValueTypes]:
+        if not self._is_answer_on_path(answer_id):
+            return None
+
+        if answer := self.answer_store.get_answer(answer_id, list_item_id):
+            return answer.value
+
+        if self.use_default_answer and (
+            answer := self.schema.get_default_answer(answer_id)
+        ):
+            return answer.value
+
     def _resolve_list_item_id_for_value_source(
         self, value_source: dict
     ) -> Optional[str]:
@@ -51,27 +72,6 @@ class ValueSourceResolver:
             else None
         )
 
-    def _is_answer_on_path(self, answer_id: str) -> bool:
-        if self.routing_path_block_ids:
-            block = self.schema.get_block_for_answer_id(answer_id)
-            return block is not None and block["id"] in self.routing_path_block_ids
-
-        return True
-
-    def _get_answer_value(
-        self, answer_id: str, list_item_id: Optional[str]
-    ) -> Optional[AnswerValueTypes]:
-        if not self._is_answer_on_path(answer_id):
-            return None
-
-        if answer := self.answer_store.get_answer(answer_id, list_item_id):
-            return answer.value
-
-        if self.use_default_answer and (
-            answer := self.schema.get_default_answer(answer_id)
-        ):
-            return answer.value
-
     def _resolve_answer_value(
         self, value_source: dict
     ) -> Union[Markup, ValueSourceTypes]:
@@ -94,19 +94,7 @@ class ValueSourceResolver:
 
         return answer_value
 
-    def _resolve_value_source_list(
-        self, value_source_list: list[dict]
-    ) -> Optional[ValueSourceTypes]:
-        values = []
-        for value_source in value_source_list:
-            value = self._resolve_value_source_dict(value_source)
-            if isinstance(value, list):
-                values.extend(value)
-            else:
-                values.append(value)
-        return values
-
-    def _resolve_value_source_dict(self, value_source: dict) -> ValueSourceTypes:
+    def resolve(self, value_source: dict) -> ValueSourceTypes:
         source = value_source["source"]
         identifier = value_source.get("identifier")
 
@@ -127,9 +115,3 @@ class ValueSourceResolver:
             # This does not use the location object because
             # routes such as individual response does not have the concept of location.
             return self.list_item_id
-
-    def resolve(self, value_source: Union[list, dict]) -> ValueSourceTypes:
-        if isinstance(value_source, list):
-            return self._resolve_value_source_list(value_source)
-
-        return self._resolve_value_source_dict(value_source)
